api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
asn1crypto==1.4.0
attrs==21.2.0
awswrangler==2.14.0
bandit==1.7.1
beautifulsoup4==4.10.0
black==21.11b1
boto3==1.20.23
botocore==1.23.23
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
coverage==6.2
cryptography==39.0.1
decorator==5.1.0
detect-secrets==1.1.0
dnspython==2.2.0
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.92.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.30
greenlet==1.1.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpath-ng==1.5.3
libsass==0.21.0
lxml==4.9.1
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.0
openpyxl==3.0.9
opensearch-py==1.0.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pbr==5.8.0
pg8000==1.22.1
platformdirs==2.4.0
playwright==1.25.2
pluggy==1.0.0
ply==3.11
progressbar2==3.55.0
psutil==5.9.2
py==1.11.0
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyee==8.1.0
pyflakes==2.4.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==7.2.0
pytest-base-url==2.0.0
pytest-cov==3.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==6.1.2
python-utils==2.5.6
pytz==2021.3
PyYAML==6.0
redshift-connector==2.0.901
regex==2021.11.10
requests==2.26.0
requests-aws4auth==1.1.1
rfc3986==1.5.0
s3transfer==0.5.0
scramp==1.4.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.25.0
stevedore==3.5.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
websockets==10.1
